+++ tests/testthat/test-check_phe_table.R
@@ -21,25 +21,12 @@ test_that("use", {
   )
 })
 
-test_that("case-control values in first column", {
+test_that("case-control values in first column is OK", {
   phe_table <- create_demo_phe_table()
   # All ones and twos
   phe_table$random <- 2
   phe_table$random[1] <- 1
-  expect_error(
-    check_phe_table(
-      phe_table
-    )
-  )
-})
-
-test_that("use, second column", {
-  phe_table <- create_demo_phe_table()
-  # All zeroes, ones and twos
-  phe_table$additive <- 2
-  phe_table$additive[1] <- 0
-  phe_table$additive[2] <- 1
-  expect_error(
+  expect_silent(
     check_phe_table(
       phe_table
     )