+++ tests/testthat/test-check_phe_table_ok_for_qt.R
@@ -1,0 +1,24 @@
+test_that("case-control values in first column", {
+  phe_table <- create_demo_phe_table()
+  # All ones and twos
+  phe_table$random <- 2
+  phe_table$random[1] <- 1
+  expect_error(
+    check_phe_table_ok_for_qt(
+      phe_table
+    )
+  )
+})
+
+test_that("use, second column", {
+  phe_table <- create_demo_phe_table()
+  # All zeroes, ones and twos
+  phe_table$additive <- 2
+  phe_table$additive[1] <- 0
+  phe_table$additive[2] <- 1
+  expect_error(
+    check_phe_table_ok_for_qt(
+      phe_table
+    )
+  )
+})